split benefit card tests
